project restructure for bundling watchface
