+++ apps/wear/smartdrive/app/App_Resources/Android/app.gradle
@@ -1,21 +1,32 @@
 // Add your native dependencies here:
 dependencies {
-  // implementation 'org.slf4j:slf4j-nop:1.7.25'
-  // implementation 'com.android.support:wear:28.0.0'
-  // implementation 'com.google.android.support:wearable:2.4.0'
-  // implementation 'com.google.api-client:google-api-client:1.28.0'
-  // implementation 'com.google.android.gms:play-services-wearable:16.0.1'
-	// implementation 'org.tensorflow:tensorflow-lite:0.0.0-nightly'
-  // compileOnly 'com.google.android.wearable:wearable:2.4.0'
+  implementation 'androidx.percentlayout:percentlayout:1.0.0'
+  implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+  implementation 'androidx.recyclerview:recyclerview:1.0.0'
+  implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+  implementation 'androidx.palette:palette:1.0.0'
 
-  implementation 'org.slf4j:slf4j-nop:1.7.25'
-  implementation 'com.android.support:wear:28.0.0'
   implementation 'com.google.android.support:wearable:2.5.0'
-  implementation 'com.google.api-client:google-api-client:1.28.0'
+  implementation 'com.android.support:wear:28.0.0'
+  implementation 'com.google.android.gms:play-services-base:17.1.0'
   implementation 'com.google.android.gms:play-services-wearable:17.0.0'
+  implementation 'com.google.api-client:google-api-client:1.28.0'
+
+  implementation 'com.github.max-mobility:circle-progress-view:2.0.1'
+
+  implementation 'io.sentry:sentry:1.7.27'
+  
+  implementation 'org.slf4j:slf4j-nop:1.7.25'
+  implementation 'org.jetbrains:annotations:16.0.2'
   implementation 'org.tensorflow:tensorflow-lite:2.0.0'
+
   compileOnly 'com.google.android.wearable:wearable:2.5.0'
+
+  // butterknife view binding https://github.com/JakeWharton/butterknife
+  implementation 'com.jakewharton:butterknife:10.2.0'
+  annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.0'
 }
+
 repositories {
   maven { url 'https://jitpack.io' }
 }
