+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-hdpi/ic_battery_full.png
