+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-hdpi/smartdrive_button.png
