+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-mdpi/ic_battery_full.png
