+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-nodpi/add_complication.png
