+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-nodpi/added_complication.png
