+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-xhdpi/ic_battery_full.png
