+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-xxhdpi/ic_battery_full.png
