+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable-xxxhdpi/ic_battery_full.png
