+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable/black_border_background.xml
