+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable/custom_complication_styles.xml
