+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable/settings_watch_face_preview_arms_and_ticks.xml
