+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/drawable/settings_watch_face_preview_background.xml
