+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/layout/activity_config.xml
