+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/layout/watchface_layout.xml
