+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/values-round/strings.xml
@@ -1,0 +1,2 @@
+<resources>
+</resources>
