+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/values/dimens.xml
