+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/res/xml/watch_face.xml
