+++ apps/wear/smartdrive/app/App_Resources/Android/src/main/resources/sentry.properties
