--- apps/wear/watchfaces/SmartDriveWatchFace/app/build.gradle
@@ -1,42 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.2"
-    defaultConfig {
-        applicationId "com.permobil.smartdrive.watchface"
-        minSdkVersion 24
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.google.android.support:wearable:2.5.0'
-    implementation 'com.google.android.gms:play-services-wearable:17.0.0'
-    implementation 'androidx.percentlayout:percentlayout:1.0.0'
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.recyclerview:recyclerview:1.0.0'
-    implementation 'com.google.android.gms:play-services-base:17.1.0'
-    implementation 'androidx.palette:palette:1.0.0'
-    implementation 'com.github.max-mobility:circle-progress-view:2.0.1'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    implementation 'org.jetbrains:annotations:16.0.2'
-    implementation 'io.sentry:sentry:1.7.27'
-    implementation 'org.slf4j:slf4j-nop:1.7.25'
-    implementation 'com.android.support:wear:28.0.0'
-
-    compileOnly 'com.google.android.wearable:wearable:2.5.0'
-
-    // butterknife view binding https://github.com/JakeWharton/butterknife
-    implementation 'com.jakewharton:butterknife:10.2.0'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.0'
-}
