--- apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/values-round/strings.xml
@@ -1,10 +1,0 @@
-<resources>
-    <string name="hello_world">Hello Round World!</string>
-    <string name="centercomplicationbackground">centerComplicationBackground</string>
-    <string name="centercomplication">centercomplication</string>
-    <string name="colon">:</string>
-    <string name="topcomplicationbackground">topcomplicationbackground</string>
-    <string name="topcomplication">topcomplication</string>
-    <string name="smartdriveimagebutton">smartDriveImageButton</string>
-    <string name="noonTime">12:00</string>
-</resources>
