--- apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/values/color.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="white">#FFFFFF</color>
-    <color name="black">#000000</color>
-    <color name="gray">#d3d3d3</color>
-
-    <!-- permobil colors -->
-    <color name="permobil_hero">#004f7e</color>
-    <color name="permobil_primary">#0067a6</color>
-    <color name="permobil_ocean">#009ac7</color>
-    <color name="permobil_cousteau">#00c1d5</color>
-    <color name="permobil_violet">#753bbd</color>
-
-    <color name="digital_date">#aaaaaa</color>
-    <color name="digital_am_pm">#aaaaa0</color>
-    <color name="digital_colons">#aaaaa0</color>
-    <color name="config_activity_background">#ffffff</color>
-    <color name="config_activity_header_text_color">#959595</color>
-    <color name="color_item_label_color">#000000</color>
-    <color name="color_item_circle_border_color">#424242</color>
-</resources>
