--- apps/wear/watchfaces/SmartDriveWatchFace/app/src/main/res/values/strings.xml
@@ -1,25 +1,0 @@
-<resources>
-    <string name="app_name">SmartDrive MX2+</string>
-    <string name="message">Watch face tapped</string>
-    <string name="my_analog_name">SD Digital</string>
-
-    <string name="digital_am">AM</string>
-    <string name="digital_pm">PM</string>
-    <string name="digital_activity_watchface_layout">DigitalWatchFaceLayoutActivity</string>
-    <!--
-    This string is used for square devices and overridden by hello_world in
-    values-round/strings.xml for round devices.
-    -->
-    <string name="smartdrivebutton">smartDriveButton</string>
-    <string name="batteryicon">batteryIcon</string>
-    <string name="centercomplicationbackground">centerComplicationBackground</string>
-    <string name="centercomplication">centercomplication</string>
-    <string name="colon">:</string>
-    <string name="topcomplicationbackground">topcomplicationbackground</string>
-    <string name="topcomplication">topcomplication</string>
-    <string name="smartdriveimagebutton">smartDriveImageButton</string>
-    <string name="noonTime">12:00</string>
-    <string name="pm">PM</string>
-    <string name="_00">00</string>
-    <string name="_12">12</string>
-</resources>
