--- apps/wear/watchfaces/SmartDriveWatchFace/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Tue Nov 12 10:11:25 CST 2019
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.6.1-all.zip
