--- apps/wear/watchfaces/SmartDriveWatchFace/local.properties
@@ -1,8 +1,0 @@
-## This file must *NOT* be checked into Version Control Systems,
-# as it contains information specific to your local configuration.
-#
-# Location of the SDK. This is only used by Gradle.
-# For customization when using a Version Control System, please read the
-# header note.
-#Tue Nov 19 13:33:57 CST 2019
-sdk.dir=/Users/bradmartin/Library/Android/sdk
