--- apps/wear/watchfaces/SmartDriveWatchFace/settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name='SmartDrive MX2+'
-include ':app'
